Update houseOccupied
Now users the person's location instead of the set tracked device for home or away. So if the user changes their device just update the person divided in settings person instead of Node-red.
Added a 1 hour pose delay to repeat all off commands to reduce irrelevant off commands being sent and processed but allows for repeat commands to ensure the first one was not missed.